Don't delete PID file on exit of asterisk -r
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@1172 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ asterisk.c
@@ -425,7 +425,7 @@ static void quit_handler(int num, int nice, int safeshutdown, int restart)
 		close(ast_consock);
 	if (ast_socket > -1)
 		unlink((char *)ast_config_AST_SOCKET);
-	unlink((char *)ast_config_AST_PID);
+	if (!option_remote) unlink((char *)ast_config_AST_PID);
 	printf(term_quit());
 	if (restart) {
 		if (option_verbose || option_console)
